# docs/openapi/v1.yaml
openapi: 3.1.0
info:
  title: LuxeKnox Gym API
  version: 0.1.0
  summary: Shared REST contract for the Nest backend and Flutter client.
  description: |
    Source of paths: `docs/backend-frd.md` §24.
    Auth and envelope: ADR-0003.
    Prefix `/v1` lives on the server URL.

    Money is a two-decimal string. IDs are int64. Timestamps are UTC ISO-8601.
    List envelope is `{ data, meta }`. Errors are `{ code, message, details, request_id }`.

    `x-status`: `module-0` (spine), `mvp` (build with the vertical), `deferred` (ADR-0005, do not call).
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: http://localhost:3000/v1
    description: Local
  - url: https://api.example.com/v1
    description: Production (replace host)
...
security:
  - bearerAuth: []
paths:
...
  /foods:
    get:
      operationId: listFoods
      summary: Food library
      tags: [DIET]
      x-status: mvp
      x-parity: required
      x-permission: diet.read
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/Q"
        - { name: is_verified, in: query, schema: { type: boolean } }
        - { name: is_active, in: query, schema: { type: boolean } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodPage"
        "400":
          $ref: "#/components/responses/Error400"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
        "404":
          $ref: "#/components/responses/Error404"
        "409":
          $ref: "#/components/responses/Error409"
        "422":
          $ref: "#/components/responses/Error422"
        "429":
          $ref: "#/components/responses/Error429"
    post:
      operationId: createFood
      summary: Create a food
      tags: [DIET]
      x-status: mvp
      x-parity: required
      x-permission: diet.create
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FoodWrite"
      responses:
        "201":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Food"
        "400":
          $ref: "#/components/responses/Error400"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
        "404":
          $ref: "#/components/responses/Error404"
        "409":
          $ref: "#/components/responses/Error409"
        "422":
          $ref: "#/components/responses/Error422"
        "429":
          $ref: "#/components/responses/Error429"
  /foods/{id}:
    get:
      operationId: getFood
      summary: Food detail
      tags: [DIET]
      x-status: mvp
      x-parity: required
      x-permission: diet.read
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Food"
        "400":
          $ref: "#/components/responses/Error400"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
        "404":
          $ref: "#/components/responses/Error404"
        "409":
          $ref: "#/components/responses/Error409"
        "422":
          $ref: "#/components/responses/Error422"
        "429":
          $ref: "#/components/responses/Error429"
    patch:
      operationId: updateFood
      summary: Update a food
      tags: [DIET]
      x-status: mvp
      x-parity: required
      x-permission: diet.update
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FoodWrite"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Food"
        "400":
          $ref: "#/components/responses/Error400"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
        "404":
          $ref: "#/components/responses/Error404"
        "409":
          $ref: "#/components/responses/Error409"
        "422":
          $ref: "#/components/responses/Error422"
        "429":
          $ref: "#/components/responses/Error429"
...
  /media/uploads:
    post:
      operationId: createMediaUpload
      summary: Signed PUT slot (ADR-0008)
      tags: [MEDIA]
      x-status: mvp
      x-parity: required
      x-permission: media.create
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MediaUploadRequest"
      responses:
        "201":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaUpload"
        "400":
          $ref: "#/components/responses/Error400"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
        "404":
          $ref: "#/components/responses/Error404"
        "409":
          $ref: "#/components/responses/Error409"
        "422":
          $ref: "#/components/responses/Error422"
        "429":
          $ref: "#/components/responses/Error429"
  /media/{key}:
    get:
      operationId: getMediaUrl
      summary: Short-lived signed GET (ADR-0008)
      tags: [MEDIA]
      x-status: mvp
      x-parity: required
      x-permission: media.read
      security:
        - bearerAuth: []
      parameters:
        - { name: key, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaDownload"
        "400":
          $ref: "#/components/responses/Error400"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
        "404":
          $ref: "#/components/responses/Error404"
        "409":
          $ref: "#/components/responses/Error409"
        "422":
          $ref: "#/components/responses/Error422"
        "429":
          $ref: "#/components/responses/Error429"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Access token `gk_at_<base64url>`. Never a JWT. Never a query param.
    deviceKey:
      type: apiKey
      in: header
      name: X-Device-Key
      description: Turnstile/reader credential. Scoped to attendance ingest only.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
      description: Default from gym_settings pagination.default_page_size.
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
      description: Admin tables that need page numbers.
...
    Q:
      name: q
      in: query
      schema:
        type: string
      description: Case-insensitive search (FR-API-014).
...
  responses:
    Error400:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    Error401:
      description: Missing or expired session
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    Error403:
      description: Permission or row-scope failure
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    Error404:
      description: Unknown or out-of-scope id
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    Error409:
      description: Version mismatch or unique conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    Error422:
      description: Business rule
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    Error429:
      description: Rate limited
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
  schemas:
    ErrorCode:
      type: string
      enum:
        - validation_error
        - unauthenticated
        - forbidden
        - not_found
        - conflict
        - business_rule
        - rate_limited
    ErrorDetail:
      type: object
      properties:
        field:
          type: string
        issue:
          type: string
    ErrorBody:
      type: object
      required: [code, message, details, request_id]
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
        details:
          type: array
          items:
            $ref: "#/components/schemas/ErrorDetail"
        request_id:
          type: string
...
    Id:
      type: integer
      format: int64
      minimum: 1
    DateTime:
      type: string
      format: date-time
      description: UTC ISO-8601
...
    PageMeta:
      type: object
      required: [limit, has_more]
      properties:
        limit:
          type: integer
        offset:
          type: integer
          nullable: true
        cursor:
          type: string
          nullable: true
        next_cursor:
          type: string
          nullable: true
        has_more:
          type: boolean
        total:
          type: integer
          nullable: true
          description: Present on offset lists; omitted on cursor feeds.
...
    Food:
      type: object
      required: [id, name, serving_unit, is_active]
      properties:
        id:
          $ref: "#/components/schemas/Id"
        name:
          type: string
        serving_unit:
          type: string
        serving_size:
          type: number
        calories:
          type: number
        protein_grams:
          type: number
        carbs_grams:
          type: number
        fat_grams:
          type: number
        fiber_grams:
          type: number
        is_verified:
          type: boolean
        is_active:
          type: boolean
    FoodWrite:
      type: object
      required: [name, serving_unit]
      properties:
        name:
          type: string
        serving_unit:
          type: string
        serving_size:
          type: number
        calories:
          type: number
        protein_grams:
          type: number
        carbs_grams:
          type: number
        fat_grams:
          type: number
        fiber_grams:
          type: number
        is_verified:
          type: boolean
        is_active:
          type: boolean
    FoodPage:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Food"
        meta:
          $ref: "#/components/schemas/PageMeta"
...
    MediaUploadRequest:
      type: object
      required: [purpose, content_type, size_bytes]
      properties:
        purpose:
          type: string
          enum: [exercise_media, avatar, progress_photo, id_proof, waiver, medical_cert, receipt_pdf]
        content_type:
          type: string
        size_bytes:
          type: integer
    MediaUpload:
      type: object
      required: [url, object_key, expires_at]
      properties:
        url:
          type: string
          format: uri
        object_key:
          type: string
        expires_at:
          $ref: "#/components/schemas/DateTime"
    MediaDownload:
      type: object
      required: [url, expires_at]
      properties:
        url:
          type: string
          format: uri
        expires_at:
          $ref: "#/components/schemas/DateTime"